pyproject.toml
[tool.poetry]
name = "exercise"
version = "0.1.0"
description = ""
authors = ["MChrys <chrys.beltran@outlook.fr>"]
readme = "README.md"



[tool.poetry.dependencies]
python = "^3.12"
aiofiles = "^24.1.0"
asyncio = "^3.4.3"
epitran = "^1.25"
langchain = "^0.2.6"
openai = "^1.35.10"
pydub = "^0.25.1"
spacy = "^3.7.5"
strsim = "^0.0.3"
streamlit = "^1.36.0"
streamlit-authenticator = "^0.3.2"
torch = "2.2.0"
torchaudio = "2.2.0"
transformers = "4.42.0"
whisperx = "3.1.1"
pyannote-audio = "3.1.1"
hydra-core = "^1.3.2"
docker = "^7.1.0"
jamo = "^0.4.1"
watchdog = "^4.0.1"
g2pk = "^0.9.4"
sentencepiece = "^0.2.0"
varname = "^0.13.1"
opentelemetry-api = "^1.25.0"
opentelemetry-sdk = "^1.25.0"
opentelemetry-instrumentation = "^0.46b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
after-build = "python -m spacy download fr_core_news_lg"

